LSR-1026 Complete VideoDisplayScreenlet: must implement IVideoDisplayScreenletListener
