+++ xamarin/Samples/Showcase-Android/Activities/MainActivity.cs
@@ -25,6 +25,7 @@ namespace ShowcaseAndroid
             FindViewById(Resource.Id.image_display_screenlet).SetOnClickListener(this);
             FindViewById(Resource.Id.pdf_display_screenlet).SetOnClickListener(this);
             FindViewById(Resource.Id.audio_display_screenlet).SetOnClickListener(this);
+            FindViewById(Resource.Id.video_display_screenlet).SetOnClickListener(this);
         }
 
         /* IOnClickListener */
@@ -65,6 +66,9 @@ namespace ShowcaseAndroid
                     break;
                 case Resource.Id.audio_display_screenlet:
                     StartActivity(typeof(AudioDisplayActivity));
+                    break;
+                case Resource.Id.video_display_screenlet:
+                    StartActivity(typeof(VideoDisplayActivity));
                     break;
             }
         }
