+++ xamarin/Samples/Showcase-Android/Resources/layout/MainView.axml
@@ -54,12 +54,12 @@
                 android:text="@string/pdf_display"
                 style="@style/test_app_button" />
             <Button
-                android:id="@+id/video_display_screenlet"
-                android:text="@string/video_display"
-                style="@style/test_app_button" />
-            <Button
                 android:id="@+id/audio_display_screenlet"
                 android:text="@string/audio_display"
+                style="@style/test_app_button" />
+            <Button
+                android:id="@+id/video_display_screenlet"
+                android:text="@string/video_display"
                 style="@style/test_app_button" />
         </LinearLayout>
     </ScrollView>
